Refactor Word Search backtracking and remove signed index casts

+++ LeetCodeCpp/WordSearch_79.h
@@ -8,5 +8,9 @@ class WordSearch_79 {
 public:
     bool exist(vector<vector<char>>& board, string word);
 private:
-    bool backtrack(vector<vector<char>>& board, const string& word, int row, int col, int index);
+    bool backtrack(vector<vector<char>>& board,
+    const string& word,
+    size_t row,
+    size_t col,
+    size_t index);
 };
